Updated contant fields, moved waiter to login method, added access modifiers to loginpage and homepage objects
