+++ Task60/src/main/java/classicpageobject/pageobject/HomePage.java
@@ -1,10 +1,10 @@
-package pageobject;
+package classicpageobject.pageobject;
 
 import org.openqa.selenium.By;
 
 public class HomePage extends Page {
     private static final String TITLE = "Входящие - Почта Mail.Ru";
-    private static final By logoutLink = By.id("PH_logoutLink");
+    private static final By LOGOUT_LINK = By.id("PH_logoutLink");
 
     public HomePage() {
         super(TITLE);
@@ -15,7 +15,7 @@ public class HomePage extends Page {
     }
 
     public LoginPage logout() {
-        driver.findElement(logoutLink).click();
+        driver.findElement(LOGOUT_LINK).click();
         return new LoginPage();
     }
