+++ Task60/src/test/java/classicpageobject/LoginTest.java
@@ -1,4 +1,4 @@
-package pageobject;
+package classicpageobject.pageobject;
 
 import driver.Driver;
 import org.openqa.selenium.WebDriver;
@@ -6,14 +6,13 @@ import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import waits.Wait;
 
 public class LoginTest {
     private static final String USERNAME = "seleniumtests10";
     private static final String PASSWORD = "060788avavav";
     private WebDriver driver;
-    LoginPage loginPage;
-    HomePage homePage;
+    private LoginPage loginPage;
+    private HomePage homePage;
 
     @BeforeMethod
     public void setup() {
@@ -29,7 +28,6 @@ public class LoginTest {
     @Test
     public void login() {
         homePage = loginPage.login(USERNAME,PASSWORD);
-        Wait.waitForTitleChange(loginPage.getTitle());
 
         Assert.assertTrue(homePage.isDisplayed(),
                 "Expected title is '" + homePage.getTitle() + "' but actual title is '" + driver.getTitle() + "'.");
@@ -38,7 +36,6 @@ public class LoginTest {
     @Test
     public void logout() {
         homePage = loginPage.login(USERNAME,PASSWORD);
-        Wait.waitForTitleChange(loginPage.getTitle());
         loginPage = homePage.logout();
 
         Assert.assertTrue(loginPage.isDisplayed(),
